BusBooking/requirements.txt
django==1.11
pillow==5.0.0
ipdb==0.11
django-extensions==2.0.3
djangorestframework==3.7.7
